refactor, split URL build from base figma session
