+++ FigmaPy/session/base.py
@@ -1,0 +1,6 @@
+class FigmaPyBase:
+    def __init__(self, token, oauth2=False):
+        self.api_uri = 'https://api.figma.com/v1/'
+        self.token_uri = 'https://www.figma.com/oauth'
+        self.api_token = token
+        self.oauth2 = oauth2
